+++ doc/node.1
@@ -140,9 +140,6 @@ Requires Node.js to be built with
 .It Fl -enable-source-maps
 Enable Source Map V3 support for stack traces.
 .
-.It Fl -experimental-global-webcrypto
-Expose the Web Crypto API on the global scope.
-.
 .It Fl -experimental-import-meta-resolve
 Enable experimental ES modules support for import.meta.resolve().
 .
@@ -165,9 +162,6 @@ Disable experimental support for the Fetch API.
 .
 .It Fl -no-experimental-global-customevent
 Disable exposition of the CustomEvent on the global scope.
-.
-.It Fl -no-experimental-global-webcrypto
-Disable exposition of the Web Crypto API on the global scope.
 .
 .It Fl -no-experimental-repl-await
 Disable top-level await keyword support in REPL.
