+++ lib/internal/bootstrap/browser.js
@@ -2,6 +2,7 @@
 
 const {
   ObjectDefineProperty,
+  ObjectGetOwnPropertyDescriptor,
   globalThis,
 } = primordials;
 
@@ -13,6 +14,10 @@ const {
   defineReplaceableLazyAttribute,
   exposeLazyInterfaces,
 } = require('internal/util');
+const {
+  ERR_INVALID_THIS,
+  ERR_NO_CRYPTO,
+} = require('internal/errors').codes;
 const config = internalBinding('config');
 
 // https://console.spec.whatwg.org/#console-namespace
@@ -143,3 +148,30 @@ exposeLazyInterfaces(
   [
     'CompressionStream', 'DecompressionStream',
   ]);
+
+// Web Crypto API
+if (config.hasOpenSSL) {
+  defineReplaceableLazyAttribute(
+    globalThis,
+    'internal/crypto/webcrypto',
+    ['crypto'],
+    false,
+    function cryptoThisCheck() {
+      if (this !== globalThis && this != null)
+        throw new ERR_INVALID_THIS(
+          'nullish or must be the global object');
+    }
+  );
+  exposeLazyInterfaces(
+    globalThis, 'internal/crypto/webcrypto',
+    ['Crypto', 'CryptoKey', 'SubtleCrypto']
+  );
+} else {
+  ObjectDefineProperty(globalThis, 'crypto',
+                       { __proto__: null, ...ObjectGetOwnPropertyDescriptor({
+                         get crypto() {
+                           throw new ERR_NO_CRYPTO();
+                         }
+                       }, 'crypto') });
+
+}
