+++ lib/internal/crypto/webcrypto.js
@@ -5,7 +5,6 @@ const {
   JSONParse,
   JSONStringify,
   ObjectDefineProperties,
-  ObjectDefineProperty,
   ReflectApply,
   ReflectConstruct,
   SafeSet,
@@ -28,10 +27,6 @@ const {
   validateOneOf,
   validateString,
 } = require('internal/validators');
-
-const {
-  getOptionValue,
-} = require('internal/options');
 
 const { TextDecoder, TextEncoder } = require('internal/encoding');
 
@@ -789,18 +784,6 @@ ObjectDefineProperties(
       value: randomUUID,
     },
   });
-
-if (getOptionValue('--no-experimental-global-webcrypto')) {
-  // For backward compatibility, keep exposing CryptoKey in the Crypto prototype
-  // when using the flag.
-  ObjectDefineProperty(Crypto.prototype, 'CryptoKey', {
-    __proto__: null,
-    enumerable: true,
-    configurable: true,
-    writable: true,
-    value: CryptoKey,
-  });
-}
 
 ObjectDefineProperties(
   SubtleCrypto.prototype, {
