+++ lib/internal/main/eval_string.js
@@ -31,9 +31,9 @@ else {
   // For backward compatibility, we want the identifier crypto to be the
   // `node:crypto` module rather than WebCrypto.
   const isUsingCryptoIdentifier =
-                             getOptionValue('--experimental-global-webcrypto') &&
-                             RegExpPrototypeExec(/\bcrypto\b/, source) !== null;
-  const shouldDefineCrypto = isUsingCryptoIdentifier && internalBinding('config').hasOpenSSL;
+    RegExpPrototypeExec(/\bcrypto\b/, source) !== null;
+  const shouldDefineCrypto = isUsingCryptoIdentifier &&
+    internalBinding('config').hasOpenSSL;
 
   if (isUsingCryptoIdentifier && !shouldDefineCrypto) {
     // This is taken from `addBuiltinLibsToObject`.
