+++ lib/internal/process/pre_execution.js
@@ -4,7 +4,6 @@ const {
   NumberParseInt,
   ObjectDefineProperties,
   ObjectDefineProperty,
-  ObjectGetOwnPropertyDescriptor,
   SafeMap,
   StringPrototypeStartsWith,
   globalThis,
@@ -18,14 +17,10 @@ const { reconnectZeroFillToggle } = require('internal/buffer');
 const {
   defineOperation,
   exposeInterface,
-  exposeLazyInterfaces,
-  defineReplaceableLazyAttribute,
 } = require('internal/util');
 
 const {
-  ERR_INVALID_THIS,
   ERR_MANIFEST_ASSERT_INTEGRITY,
-  ERR_NO_CRYPTO,
 } = require('internal/errors').codes;
 const assert = require('internal/assert');
 const {
@@ -64,7 +59,6 @@ function prepareExecution(options) {
   setupInspectorHooks();
   setupWarningHandler();
   setupFetch();
-  setupWebCrypto();
   setupCustomEvent();
 
   // Resolve the coverage directory to an absolute path, and
@@ -267,41 +261,6 @@ function setupFetch() {
   internalBinding('wasm_web_api').setImplementation((streamState, source) => {
     require('internal/wasm_web_api').wasmStreamingCallback(streamState, source);
   });
-}
-
-// TODO(aduh95): move this to internal/bootstrap/browser when the CLI flag is
-//               removed.
-function setupWebCrypto() {
-  if (process.config.variables.node_no_browser_globals ||
-      getOptionValue('--no-experimental-global-webcrypto')) {
-    return;
-  }
-
-  if (internalBinding('config').hasOpenSSL) {
-    defineReplaceableLazyAttribute(
-      globalThis,
-      'internal/crypto/webcrypto',
-      ['crypto'],
-      false,
-      function cryptoThisCheck() {
-        if (this !== globalThis && this != null)
-          throw new ERR_INVALID_THIS(
-            'nullish or must be the global object');
-      }
-    );
-    exposeLazyInterfaces(
-      globalThis, 'internal/crypto/webcrypto',
-      ['Crypto', 'CryptoKey', 'SubtleCrypto']
-    );
-  } else {
-    ObjectDefineProperty(globalThis, 'crypto',
-                         { __proto__: null, ...ObjectGetOwnPropertyDescriptor({
-                           get crypto() {
-                             throw new ERR_NO_CRYPTO();
-                           }
-                         }, 'crypto') });
-
-  }
 }
 
 // TODO(daeyeon): move this to internal/bootstrap/browser when the CLI flag is
