+++ test/parallel/test-cli-eval.js
@@ -348,12 +348,6 @@ child.exec(
   common.mustSucceed((stdout) => {
     assert.match(stdout, /^number/);
   }));
-child.exec(
-  `${nodejs} --no-experimental-global-webcrypto ` +
-    '-p "var crypto = {randomBytes:1};typeof crypto.randomBytes"',
-  common.mustSucceed((stdout) => {
-    assert.match(stdout, /^number/);
-  }));
 
 // Regression test for https://github.com/nodejs/node/issues/45336
 child.execFile(process.execPath,
